Notifications now have a separate repository
