+++ src/Chirp.Core/ICheepRepository.cs
@@ -6,9 +6,5 @@ public interface ICheepRepository
     public void UpdateMessage();
     public Task<List<CheepDTO>> GetMessagesFromAuthor(string author, int page);
     public bool CreateCheep(AuthorDTO author, string text, List<AuthorDTO>? mentions, string time);
-    public Task<List<NotificationDTO>> GetNotifications(string username);
-    public Task DeleteNotification(int notificationId);
-    public Task ForgetMentions(string authorName);
-    public Task DeleteNotificationsForUser(int userId);
     Task<CheepDTO> GetCheepById(int id);
 }
