+++ src/Chirp.Core/INotificationsRepository.cs
@@ -1,0 +1,9 @@
+namespace Chirp.Core;
+
+public interface INotificationRepository
+{
+    public Task<List<NotificationDTO>> GetNotifications(string username);
+    public Task DeleteNotification(int notificationId);
+    public Task ForgetMentions(string authorName);
+    public Task DeleteNotificationsForUser(int userId);
+}
