+++ src/Chirp.Web/Pages/ForgetMe.cshtml.cs
@@ -10,13 +10,13 @@ namespace Chirp.Web.Pages;
 public class ForgetMe : PageModel
 {
     private IAuthorRepository _repository;
-    private ICheepRepository _cRepository;
+    private INotificationRepository _nRepository;
     private SignInManager<ChirpUser> _signInManager;
     
-    public ForgetMe(IAuthorRepository repository, ICheepRepository cheepRepository, SignInManager<ChirpUser> signInManager)
+    public ForgetMe(IAuthorRepository repository, INotificationRepository notificationRepository, SignInManager<ChirpUser> signInManager)
     {
         _repository = repository;
-        _cRepository = cheepRepository;
+        _nRepository = notificationRepository;
         _signInManager = signInManager;
     }
     
@@ -25,8 +25,8 @@ public class ForgetMe : PageModel
         var username = User.FindFirstValue(ClaimTypes.Name);
         var email = User.FindFirstValue(ClaimTypes.Email);
         if(username != null && email != null) {
-            _cRepository.ForgetMentions(username);
-            _cRepository.DeleteNotificationsForUser(_repository.GetAuthor(username, email).Id);
+            _nRepository.ForgetMentions(username);
+            _nRepository.DeleteNotificationsForUser(_repository.GetAuthor(username, email).Id);
             _repository.DeleteUser(username,email);
             _signInManager.SignOutAsync();
