+++ src/Chirp.Web/Program.cs
@@ -45,6 +45,7 @@ builder.Services.AddSession(options =>
 // Add services to the container.
 builder.Services.AddScoped<IAuthorRepository, AuthorRepository>();
 builder.Services.AddScoped<ICheepRepository, CheepRepository>();
+builder.Services.AddScoped<INotificationRepository, NotificationRepository>();
 builder.Services.AddRazorPages();
